chore!: Change attributes for `tool-name` and `tool-title`
`title` is a reserved attribute in HTML and lit elements,
so we need to avoid using it. I considered also changing
`tool-icon` (and maybe also `tool-intent`), but decided that
the minimal change to have tool name and title is enough.
If we only change `tool-title`, though, then it's odd that
`name` is not `tool-name` as well.