fix: Send a `shiny-tool-request-hide` custom message on result
This is required when the tool result display is completely customized by the user, i.e they don't use our `<shiny-tool-result>` component.
Without the custom message, we can't hide the tool result.

Interestingly we need to remember the message because the tool
request and result are being re-rendered frequently while the LLM is
streaming output. For now, we store a set of seen tool request IDs
and we short-circuit tool request rendering if we've already seen
that ID. In the future, if we start new markdown-stream components
when we have a switch in content type, we won't need to remember
between renders.